:recycle: move cookies
