+++ src/handlers.rs
@@ -1,4 +1,4 @@
 pub mod auth;
-pub mod cookies;
-pub mod features_i_tried;
 pub mod tasks;
+
+pub mod features_i_tried;
