+++ src/main.rs
@@ -23,7 +23,6 @@ mod models;
 mod schema;
 
 mod handlers;
-use handlers::cookies;
 
 #[launch]
 fn rocket() -> Rocket<Build> {
@@ -36,7 +35,6 @@ fn rocket() -> Rocket<Build> {
         .attach(fairing::log::LoggingUidFairing)
         .attach(fairing::migration::MigrationFairing)
         .attach(handlers::features_i_tried::stage())
-        .mount("/", routes![cookies::cookies, cookies::set_cookies,])
         .mount("/public", FileServer::from(&CONFIG.public_dir))
         .attach(handlers::auth::stage())
         .attach(handlers::tasks::stage())
